Disable Harfbuzz/CoreText hotfix on older macOS/iOS versions
Since the new default font on macOS 11 and iOS 14 does not have the
AAT tables we usually check for, we added a check for TRAK as well,
otherwise we would end up applying the wrong kerning rules to the
font in these newer versions of the operating systems.

However, this introduced a regression on older versions when letter
spacing was set. This basically just disables the ligature features
when shaping, so the reason is not fully understood, but the
assumption is that it exposes a bug in the CoreText shaper backend
in Harfbuzz which was previously hidden.

The pragmatic fix for this is a second hotfix which disables the
previous one on older OS versions, since:
  A. The original bug does not occur on older versions.
  B. Newer versions of Harfbuzz (and therefore Qt) does not exhibit
     the problem since this code path is no longer used.

To do this we also need to rename the hb-coretext.cc file to .mm
to enable using the Objective-C APIs.

This is all pretty ugly, but it is not upstreamed anywhere and has
limited lifespan.

[ChangeLog][macOS/iOS] Fixed a regression when combining letter
spacing with the default font on macOS 10.15 and lower or iOS 13
or lower.

Fixes: QTBUG-91764
Change-Id: I50c8ad08d51899635418aeaa38230bbeab5c5ef7
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>